feat(release): added semantic-release
