+++ .storybook/webpack.config.js
@@ -90,7 +90,7 @@ module.exports = async ({ config, mode }) => {
   });
 
   config.module.rules.push({
-    test: /styles.scss$/,
+    test: /storybook.scss$/,
     use: ['style-loader', 'css-loader', 'sass-loader'],
     include: path.resolve(__dirname, '../'),
   });
