+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@wfp/ui",
-  "version": "0.0.0",
+  "version": "0.0.0-development",
   "description": "WFP UI Kit Next",
   "license": "Apache-2",
   "main": "lib/index.js",
