+++ src/components/Link/Link.js
@@ -4,12 +4,8 @@ import classnames from 'classnames';
 import styles from './link.module.scss';
 
 /**
-Links are typically used as a means of navigation either within the application, to a place outside, or to a resource. For anything else, especially things that change data and actions, you should be using a button.
-
+Links are typically used as a means of navigation either within the application, to a place outside, or to a resource. For anything else, especially things that change data and actions, you should be using a `button`.
 Links can have the same properties as a regular `<a>`-element.
-
-Links is based on a fork from [Carbon Components](https://www.carbondesignsystem.com/components/link/code) 
-
 
 ### Content
 - Use text for links rather than graphics or icons.
