+++ src/components/Link/README.md
@@ -1,4 +1,4 @@
-Hello Links are typically used as a means of navigation either within the application, to a place outside, or to a resource. For anything else, especially things that change data and actions, you should be using a button.
+Links are typically used as a means of navigation either within the application, to a place outside, or to a resource. For anything else, especially things that change data and actions, you should be using a button.
 
 Links can have the same properties as a regular `<a>`-element.
