+++ src/documentation/Logos/Logos.stories.mdx
@@ -1,0 +1,12 @@
+import { Meta, Story, Preview } from '@storybook/addon-docs/blocks';
+import Button from '../../components/Button';
+import Logos from './Logos';
+
+<Meta title="Documentation/Logo" />
+
+# Checkbox
+
+With `MDX` we can define a story for `Checkbox` right in the middle of our
+markdown documentation.
+
+<Logos />
